Fix Lease status transition; Add Lease Metadata
Add metadata field support to the POST /leases endpoint

Fix a bug where if a principal was assigned to an account they'd
previously used, the lease would be marked as active, but no other
fields would be updated. So, for example, the expiredOn field would
remain in the past, metadata wouldn't update, etc.

Also:
- Remove Provisioner pkg (no longer used, with new UpsertLease logic)
- Fix references in code/logs/tests to "Provisioning / Decommissioning".
  Preferred language is "create/destroy lease"
- A few logging improvements
- A few var renames